start fixing auth/profile tests
